Project/requirements.txt
tensorflow==2.9.1
numpy==1.23.5
matplotlib==3.5.2
flask==2.1.2
pillow==9.2.0